Split net files
