+++ src/sys/net/nic/mod.rs
@@ -1,0 +1,2 @@
+pub mod rtl8139;
+pub mod pcnet;
